+++ release.bash
@@ -6,4 +6,4 @@ rustfmt --check src/**/*.rs &&
     cargo test &&
     cargo build --release &&
     mkdir -p ${RELEASE_DIR} &&
-    cp -r CHANGELOG.md commands.md example/ img/ README.md LICENSE target/release/rtw ${RELEASE_DIR}
+    cp -r ./CHANGELOG.md ./commands.md ./shell-completion.md example/ img/ ./README.md ./LICENSE target/release/rtw ${RELEASE_DIR}