+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.5.0](https://crates.io/crates/rtw/1.5.0) Jun 21, 2020 
+
+* add `completion <shell>` command.
+
+`rtw completion <shell>` generates completion file for `<shell>`
+
 ## [1.4.1](https://crates.io/crates/rtw/1.4.1) Jun 19, 2020 
 
 * Fix timeline crash when activity spans over several days (#33)